Split auth fail test into separate tests.
